Remove redundant u64 cast in prepare_tree windowed sync
completing_block_height is already u64; clippy::unnecessary_cast fails CI with -D warnings.

Made-with: Cursor

+++ witness/witness-server/src/server.rs
@@ -136,7 +136,7 @@ async fn prepare_tree(
             })
             .collect();
 
-        let sync_from = subtree_roots[window_start - 1].completing_block_height as u64 + 1;
+        let sync_from = subtree_roots[window_start - 1].completing_block_height + 1;
         let initial_tree_size = Some(leaf_offset as u32);
 
         tracing::info!(
